+++ src/r0/sync/get_message_events.rs
@@ -6,6 +6,8 @@ use ruma_events::{collections::only, EventResult};
 use ruma_identifiers::RoomId;
 use serde::{Deserialize, Serialize};
 
+use crate::r0::filter::RoomEventFilter;
+
 ruma_api! {
     metadata {
         description: "Get message events for a room.",
@@ -44,6 +46,10 @@ ruma_api! {
         #[serde(skip_serializing_if = "Option::is_none")]
         #[ruma_api(query)]
         pub limit: Option<UInt>,
+        /// A RoomEventFilter to filter returned events with.
+        #[serde(skip_serializing_if = "Option::is_none")]
+        #[ruma_api(query)]
+        pub filter: Option<RoomEventFilter>,
     }
 
     response {